Extract reset password flow into its own module
